# add optimized load_raw_field()