... for a 100x speed-up.

Old version:
```julia
julia> import SloanDigitalSkySurvey.SDSS

julia> @time SDSS.load_raw_field(".", "003900", "6", "0269", 1, 1.);
  8.682727 seconds (71.00 M allocations: 1.636 GB, 28.46% gc time)

julia> @time SDSS.load_raw_field(".", "003900", "6", "0269", 1, 1.);
  7.143706 seconds (68.65 M allocations: 1.535 GB, 43.42% gc time)
```

New version:
```julia
julia> import Celeste.SkyImages

julia> @time SkyImages.load_raw_field(".", "003900", "6", "0269", 1, 1.);
  0.141113 seconds (87.34 k allocations: 27.344 MB)

julia> @time SkyImages.load_raw_field(".", "003900", "6", "0269", 1, 1.);
  0.056292 seconds (187 allocations: 23.497 MB)
```

Still need to write a test for `interp_sky()`. Hold off on merging until I push that, probably tomorrow.