@jeff-regier 2016-02-15
Nice work Kyle! 100x is a beautiful thing...

On Mon, Feb 15, 2016, 3:14 AM Kyle Barbary <notifications@github.com> wrote:

> ... for a 100x speed-up.
>
> Old version:
>
> julia> import SloanDigitalSkySurvey.SDSS
>
> julia> @time SDSS.load_raw_field(".", "003900", "6", "0269", 1, 1.);
>   8.682727 seconds (71.00 M allocations: 1.636 GB, 28.46% gc time)
>
> julia> @time SDSS.load_raw_field(".", "003900", "6", "0269", 1, 1.);
>   7.143706 seconds (68.65 M allocations: 1.535 GB, 43.42% gc time)
>
> New version:
>
> julia> import Celeste.SkyImages
>
> julia> @time SkyImages.load_raw_field(".", "003900", "6", "0269", 1, 1.);
>   0.141113 seconds (87.34 k allocations: 27.344 MB)
>
> julia> @time SkyImages.load_raw_field(".", "003900", "6", "0269", 1, 1.);
>   0.056292 seconds (187 allocations: 23.497 MB)
>
> Still need to write a test for interp_sky(). Hold off on merging until I
> push that, probably tomorrow.
> ------------------------------
> You can view, comment on, or merge this pull request online at:
>
>   https://github.com/jeff-regier/Celeste.jl/pull/119
> Commit Summary
>
>    - add optimized load_raw_field()
>
> File Changes
>
>    - *M* src/SkyImages.jl
>    <https://github.com/jeff-regier/Celeste.jl/pull/119/files#diff-0> (88)
>
> Patch Links:
>
>    - https://github.com/jeff-regier/Celeste.jl/pull/119.patch
>    - https://github.com/jeff-regier/Celeste.jl/pull/119.diff
>
> —
> Reply to this email directly or view it on GitHub
> <https://github.com/jeff-regier/Celeste.jl/pull/119>.
>